requirements.txt
Pillow
protobuf
streamlit
google-generativeai
PyPDF2
gspread